requirements.txt
python-dotenv==1.0.0
langchain==0.3.13
langchain-core==0.3.27
langchain-community==0.3.13
fastapi==0.109.0
uvicorn[standard]==0.27.0.post1
openai==1.58.1
langchain-openai==0.2.13
termcolor==2.4.0
python-multipart==0.0.20
pypdf==5.1.0